pyproject.toml
[project]
name = "rag-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Rishav",email = "rishishah907@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "weaviate-client (>=4.14.3,<5.0.0)",
    "python-docx (>=1.1.2,<2.0.0)",
    "pymupdf (>=1.25.5,<2.0.0)",
    "pytesseract (>=0.3.13,<0.4.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "openai (>=1.81.0,<2.0.0)",
    "zappa (>=0.60.1,<0.61.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "pydantic[email] (>=2.11.4,<3.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "mangum (>=0.19.0,<0.20.0)",
    "pre-commit (>=4.2.0,<5.0.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
